+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Remove bad paths from LC-MS directory schema
 - Allow multiple comma-separated parent_sample_id values
 - Accommodate dir schema minor versions
+- Fix ORCID URL checking
 - Updating documentation
 
 ## v0.0.18